fix warning

+++ test/aead.cpp
@@ -75,7 +75,7 @@ void aead_encrypt_test(
     cout << endl;
 
     cout << "ctxt: ";
-    for (int i = 0; i < msg_len; i++) {
+    for (size_t i = 0; i < msg_len; i++) {
         cout << hex << (int)ctxt[i];
     }
     cout << endl;
@@ -141,7 +141,7 @@ void aead_decrypt_test(
     if (party == ALICE) {
         cout << "ALICE res: " << res << endl;
         cout << "ALICE msg: ";
-        for (int i = 0; i < msg_len; i++) {
+        for (size_t i = 0; i < msg_len; i++) {
             cout << hex << (int)msg[i];
         }
         cout << endl;
@@ -149,7 +149,7 @@ void aead_decrypt_test(
     } else {
         cout << "BOB res: " << res << endl;
         cout << "BOB msg: ";
-        for (int i = 0; i < msg_len; i++) {
+        for (size_t i = 0; i < msg_len; i++) {
             cout << hex << (int)msg[i];
         }
         cout << endl;
